Fix : sideProfile Toggle made to component
